BUGFIX: Remove crashes by removing potential overflows in calculation.
We think that overflow in integer calculations (squaring the radius)
has caused both errors and crashes, especially on windows for some
reason.

So this patch removes all the cases I could find by either extending
the integer range by using qint64 or converting to double, whichever
makes most sense in each case.


svn path=/branches/KDE/4.0/kdeedu/marble/; revision=776347

+++ src/lib/GlobeScanlineTextureMapper.cpp
@@ -78,8 +78,8 @@ void GlobeScanlineTextureMapper::resizeMap(int width, int height)
 
 void GlobeScanlineTextureMapper::mapTexture( ViewParams *viewParams )
 {
-    QImage* canvasImage = viewParams->m_canvasImage;
-    const int radius = viewParams->m_radius;
+    QImage       *canvasImage = viewParams->m_canvasImage;
+    const qint64  radius      = viewParams->m_radius;
 
     // Scanline based algorithm to texture map a sphere
